[REF] mail: add mail.notify.mixin
Aims to support notification for records without inheriting
the mail.thread mixin.

Task-3586524
